requirement.txt
pip -- 23.3.2
discord.py -- 2.3.2
wavelink -- 3.1.0
pytest 7.4.2

